env: do not recursively init XiangShan
XiangShan contains multiple copies of rocket-chip, huancun, and other
submodules in its submodule. If we execute `git submodule update --init
--recursive` on XiangShan, we will get multiple copies of these modules and
even have different versions. But in the mill file, we will only choose one
to compile. This behavior will confuse the developers, and they need to
exclude these modules when doing a full-text search on the XiangShan
folder. But we can get this solved by not recursively init the XiangShan
submodule.